+++ client/public/index.html
@@ -5,17 +5,10 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.3/css/all.css" integrity="sha384-SZXxX4whJ79/gErwcOYf+zWLeJdY/qpuqC4cAa9rOGUstPomtqpuNWT9wdPEn2fk" crossorigin="anonymous">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-<link href="https://fonts.googleapis.com/css2?family=Trade+Winds&display=swap" rel="stylesheet">
-<link rel="preconnect" href="https://fonts.gstatic.com">
-<link href="https://fonts.googleapis.com/css2?family=Grenze+Gotisch:wght@300;400&family=Iceberg&family=Megrim&display=swap" rel="stylesheet">
-
-<link rel="icon" type="image/png" href="tt.png"/>
-<link rel="icon" type="image/png" href="tt.png"/>
+    <link rel="icon" type="image/png" href="%PUBLIC_URL%/tt.png"/>
     <meta
-      name="description"
-      content="Web site created using create-react-app"
+      name="Tabletop Tracker"
+      content="Web app that enables users to browse and search for board games"
     />
 
     <title>Tabletop Tracker</title>
