+++ client/src/App.scss
@@ -1,6 +1,4 @@
-// $button-color: #f3f4f6;
 $button-color: #f0b214;
-// $button-border: rgba(27, 31, 35, 0.15);
 $button-border: #ce970d;
 
 html {
